move CAPD packages to internal
