+++ docs/book/src/developer/providers/v1.0-to-v1.1.md
@@ -41,6 +41,12 @@ are kept in sync with the versions used by `sigs.k8s.io/controller-runtime`.
     * [test/infrastructure/docker/controllers](https://github.com/kubernetes-sigs/cluster-api/pull/5595) 
     * [test/infrastructure/docker/exp/controllers](https://github.com/kubernetes-sigs/cluster-api/pull/5690) 
     * [exp/controllers](https://github.com/kubernetes-sigs/cluster-api/pull/5651)
+  
+* Following packages have been moved to internal
+    * [/test/infrastructure/docker/cloudinit](https://github.com/kubernetes-sigs/cluster-api/pull/5795)
+    * [/test/infrastructure/docker/docker](https://github.com/kubernetes-sigs/cluster-api/pull/5795)
+    * [/test/infrastructure/docker/third_party](https://github.com/kubernetes-sigs/cluster-api/pull/5795)
+    * [/test/infrastructure/docker/exp/docker](https://github.com/kubernetes-sigs/cluster-api/pull/5795)
 
 ### Other
