+++ test/infrastructure/docker/internal/cloudinit/doc.go
