+++ test/infrastructure/docker/internal/cloudinit/kindadapter.go
