+++ test/infrastructure/docker/internal/cloudinit/kindadapter_test.go
