+++ test/infrastructure/docker/internal/cloudinit/unknown.go
