+++ test/infrastructure/docker/internal/cloudinit/writefiles.go
