+++ test/infrastructure/docker/internal/cloudinit/writefiles_test.go
