+++ test/infrastructure/docker/internal/docker/types/node.go
